requirements.txt
flask==3.0.2
pandas==2.2.1
langchain-groq==0.1.0
langchain-core==0.1.27
python-dotenv==1.0.1
opencv-python==4.9.0.80
pytesseract==0.3.10
pydantic==2.6.3
requests==2.31.0
faker==22.6.0 
pymongo